+++ pre-cmake/tools/f2pmchart.cpp
@@ -6,7 +6,7 @@
 
 #include <iostream>
 #include "f2polymod_t.h"
-#include "f2poly_factor.h"
+#include "f2_poly_factor.h"
 #include "f2polymod_units.h"
 #include "minimal_and_characteristic_polynomials.h"
 #include "moebius.h"
