+++ src/factorization/f2_poly_factor.h
@@ -9,7 +9,7 @@
 #include "f2poly_t.h"
 #include "tfacinfo.h"
 
-tfacinfo<f2poly_t> f2poly_factor(f2poly_t f);
+tfacinfo<f2poly_t> f2_poly_factor(f2poly_t f);
 
 int f2poly_is_irreducible(f2poly_t f);
