+++ src/factorization/fp_poly_totient.cpp
@@ -35,22 +35,21 @@
 
 // which recovers the special case for finite fields.
 
-#include "f2npoly_totient.h"
-#include "f2npoly_factor.h"
+#include "fp_poly_totient.h"
+#include "fp_poly_factor.h"
 #include "int_power.h"
 
-int f2npoly_totient(f2npoly_t f) {
+int fp_poly_totient(fppoly_t f) {
   int rv = 1;
-  f2poly_t m = f.get_coeff(0).get_modulus();
-  int q = 1 << m.find_degree();
-  tfacinfo<f2npoly_t> finfo = f2npoly_factor(f);
+  int p = f.get_char();
+  tfacinfo<fppoly_t> finfo = fp_poly_factor(f);
   int nf = finfo.get_num_distinct();
 
   for (int i = 0; i < nf; i++) {
-    f2npoly_t fi = finfo.get_ith_factor(i);
+    fppoly_t fi = finfo.get_ith_factor(i);
     int ei = finfo.get_ith_count(i);
     int di = fi.find_degree();
-    rv *= int_power(q, di * (ei - 1)) * (int_power(q, di) - 1);
+    rv *= int_power(p, di * (ei - 1)) * (int_power(p, di) - 1);
   }
 
   return rv;
